move files to uniform stack architecture format
